changed the imageUrl to imageUrl? and move the UserProps to the file, changing the name to User only
